Separated dev and prod middlewares (#1556)
