+++ server/argv.js
@@ -1,0 +1,1 @@
+module.exports = require('minimist')(process.argv.slice(2));
