+++ server/index.js
@@ -3,7 +3,8 @@
 const express = require('express');
 const logger = require('./logger');
 
-const argv = require('minimist')(process.argv.slice(2));
+const argv = require('./argv');
+const port = require('./port');
 const setup = require('./middlewares/frontendMiddleware');
 const isDev = process.env.NODE_ENV !== 'production';
 const ngrok = (isDev && process.env.ENABLE_TUNNEL) || argv.tunnel ? require('ngrok') : false;
@@ -23,8 +24,6 @@ setup(app, {
 const customHost = argv.host || process.env.HOST;
 const host = customHost || null; // Let http.Server use its default IPv6/4 host
 const prettyHost = customHost || 'localhost';
-
-const port = argv.port || process.env.PORT || 3000;
 
 // Start your app.
 app.listen(port, host, (err) => {
