+++ server/middlewares/dllPlugin.js
@@ -1,0 +1,3 @@
+  const path = require('path');
+  const pkg = require(path.resolve(process.cwd(), 'package.json'));
+  module.exports = pkg.dllPlugin;
