+++ server/port.js
@@ -1,0 +1,3 @@
+const argv = require('./argv');
+
+module.exports = parseInt(argv.port || process.env.PORT || '3000', 10);
